+++ src/main/java/com/atlauncher/data/Instance.java
@@ -1758,11 +1758,6 @@ public class Instance extends MinecraftVersion {
             return false;
         }
 
-        if (launcher.loaderVersion.isNeoForge()) {
-            LogManager.debug("Instance " + launcher.name + " cannot be exported due to: NeoForge is not supported");
-            return false;
-        }
-
         return true;
     }